Renamed score to core and fixed some broken test cases
