+++ simbad/core/__init__.py
